e_commerce/requirements.txt
Django
pandas
django-cors-headers
djangorestframework
djangorestframework-simplejwt